+++ adminalerts/tests/test_permissions.py
@@ -1,11 +1,11 @@
 """Tests for permissions in the adminalerts app"""
-from django.http import HttpResponseBadRequest
+
 from django.urls import reverse
 
 # Projectroles dependency
 from projectroles.tests.test_permissions import TestPermissionBase
 
-from .test_models import AdminAlertMixin
+from adminalerts.tests.test_models import AdminAlertMixin
 
 
 class TestAdminAlertPermissions(AdminAlertMixin, TestPermissionBase):
@@ -75,14 +75,4 @@ class TestAdminAlertPermissions(AdminAlertMixin, TestPermissionBase):
         good_users = [self.superuser, self.regular_user]
         bad_users = [self.anonymous]
         self.assert_response(url, good_users, 200)
-        self.assert_response(url, bad_users, 302)
-
-    def test_alert_activation(self):
-        """Test permissions for AdminAlert activation API view"""
-        url = reverse('adminalerts:api_alert_activation')
-        good_users = [self.superuser]
-        bad_users = [self.anonymous, self.regular_user]
-        self.assert_response(
-            url, good_users, HttpResponseBadRequest.status_code, method='POST'
-        )
         self.assert_response(url, bad_users, 302)