+++ adminalerts/urls.py
@@ -1,10 +1,11 @@
 from django.conf.urls import url
 
+import adminalerts.views_ajax
 from . import views
 
 app_name = 'adminalerts'
 
-urlpatterns = [
+urls_ui = [
     url(regex=r'^list$', view=views.AdminAlertListView.as_view(), name='list'),
     url(
         regex=r'^detail/(?P<uuid>[0-9a-f-]+)$',
@@ -26,9 +27,14 @@ urlpatterns = [
         view=views.AdminAlertDeleteView.as_view(),
         name='delete',
     ),
+]
+
+urls_ajax = [
     url(
-        regex=r'^api/update-state',
-        view=views.AdminAlertActivationAPIView.as_view(),
-        name='api_alert_activation',
+        regex=r'^ajax/active/toggle/(?P<adminalert>[0-9a-f-]+)',
+        view=adminalerts.views_ajax.AdminAlertActiveToggleAjaxView.as_view(),
+        name='ajax_active_toggle',
     ),
 ]
+
+urlpatterns = urls_ui + urls_ajax